iot-gateway/requirements.txt
paho-mqtt==1.5.1
PyJWT==2.1.0
cryptography==3.4.7
